LIBIMOBILEDEVICE_API definition: support Visual C++
On the Win32 platform, define LIBIMOBILEDEVICE_API in the public header.
This fixes Visual C++ compiler errors; the Visual C++ compiler requires
the defintion of the method in the .h and .c files to be exactely the
same.

When compiling on other platforms, the value of the
LIBIMOBILEDEVICE_API definition depends on config.h and
cannot be defined in public headers; so for these platforms the definition
remains in idevice.h