Function declarations in public headers and source files match on msvc
Visual C++ requires the function declarations in the public headers and
source files to be exactely the same. Other compilers don't.

When compiling with MSVC, the methods declarations in the public header
are prefixed with __declspec( dllexport ) through the use of a
LIBIMOBILEDEVICE_API_WIN32 preprocessor variable. In the source files we
keep the LIBIMOBILEDEVICE_API preprocessor variable.

Using one single preprocessor variable is not possible, because you cannot
reliably determine the value of the LIBIMOBILEDEVICE_API preprocessor
variable in the public header files.